Added a way to render debug lines.
Creates a new line manager that keeps track of the mesh and points.
RenderMesh now has a primitive enum that tells the backend what
primitiveprimitive we're using.
Added a context field that commands can use to give primitive-dependent
values to the renderer. This is used to pass in the glLineWidth value.