[rothko][emulator] Create all textures at one.
This creates all the necessary textures to composite the scene.
Filled the sprite one with transparent pattern (for display purposes). A new
"real" sprites texture will have to be created that has the actual
transparent pixels.

Changed texture.dims to texture.size
Changed the signature of RendererSubTexture to accept defaults that
permits to upload the whole texture without redundant arguments.