[rothko][animation] Added ozz third party library.
Ozz is an animation library that does animation loading, evaluation and
even skinning (on CPU).

This brings the "bundled" source that can easily be compiled from within
our own build system, without us needing to do any CMake integration or
bring in a shared library (SDL-style).

Added the beginning of an example that will use the library as a test.

Added some LICENSE updates.