[rothko][scene_graph] Added initial scene graph support.
Created the plumbing, but nothing is really using it.

Started an example application that will run it. While doing it, based
on the original Cube example, started cleaning a lot of bits so that the
default case becomes more usable: Things like updating the default
camera, creating a default cube mesh with a default shader.

Fixed some compilation issues on linux (mostly GN issues).