[rothko][gltf] Updated texture/material mode to match GLTF.
GLTF has some opinions about how materials/textures are layout and
modified rothko's implementation to match that.

Moved the texture modes information (filter, wrap, etc.) to the texture
instead of using a config struct. This is simpler and also permits the
system to query that information (before it was lost).