[rothko] Correct widget for scene graph.
Will take in account the parent transformation and inverse it before
calculating the delta. This permits to create changes that propagate
correctly underneath the graph tree, regardless of the transformations
of the parents.

Fix a bug with the adjunct matrix calculation, which was being calculated
transposed (the test was a simetric matrix). Added better tests for both
adjunct matrix and inverse matrix.