[rothko] Create light widget manager.
This permits to get widgets for each type of light, without having to
create meshes and render commands directly. Supports points light for now.

Updated simple_lighting and textured_lighting examples to use this.

Fixed a bug in textured_lighting where the UBO was not passing the
normal matrix correctly due to the mat3 alignment.