[rothko][textured_lighting] Directional light.
Implemented directional lights by multiplexing the light position.
If w == 0, it is treated as a direction. Otherwise, it will do the
normal point light position calculation.

Added a no-op render command so that some widget creators (like the
LineManager) can send a no-op as a command instead of the user having
to check for conditions.