[rothko] Arbitrary UBO defnitions.
Before rothko was harcoded to expect one vertex shader UBO and a
fragment shader one. With this new definition, a shader can define an
arbitrary amount (up to a defined max) in any order between vertex and
fragment shader. This makes it easy to reuse constant values (such as
overall model transform) and not have many redundant copies of it.

This is used in the glTF example to handle the model/model_node
transforms.